refactor: put user module in user folder
