+++ src/app/modules/user/user.controller.ts
@@ -1,8 +1,8 @@
 import { User } from '@prisma/client';
 import { Request, Response } from 'express';
 import { UserService } from './user.service';
-import catchAsync from '../../shared/catch-async';
-import sendResponse from '../../shared/send-response';
+import catchAsync from '../../../shared/catch-async';
+import sendResponse from '../../../shared/send-response';
 
 const createUser = catchAsync(async (req: Request, res: Response) => {
   const data = await UserService.createUser(req.body);
